Addressing review comments:
* don't log an error in the methods; propagate err to main and log there
* external-dns code to its own file
* register provider in main() instead of init()
